Update ValidParinthesis.java
When encountering the '*' character in the input string, we have three options according to the problem statement:

Treat '*' as '(' (left parenthesis).
Treat '*' as ')' (right parenthesis).
Treat '*' as an empty string (meaning it doesn't contribute to the balance of parentheses).
In this section of the code:

leftMin--: We decrement leftMin because we are treating '' as if it were '(' (left parenthesis). This indicates that we have one less '(' available, as '' has been used to balance it.
leftMax++: We increment leftMax because we are treating '*' as if it were ')' (right parenthesis). This indicates that we potentially have one more ')' available if needed.
By doing this, we're exploring all possible scenarios that the '*' character can represent - either as '(' or ')' - and considering how it affects the minimum and maximum counts of '(' characters needed to balance the string.